requirements-py26.txt
pyzmq
M2Crypto
httplib2
argparse
msgpack-python>=0.4
psutil
ipy
ordereddict